+++ src/AzzyBot-Next/Migrations/20240715112456_InitialCreation.Designer.cs
@@ -12,7 +12,7 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace AzzyBot.Migrations
 {
     [DbContext(typeof(AzzyDbContext))]
-    [Migration("20240706180132_InitialCreation")]
+    [Migration("20240715112456_InitialCreation")]
     partial class InitialCreation
     {
         /// <inheritdoc />
@@ -20,7 +20,7 @@ namespace AzzyBot.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "8.0.6")
+                .HasAnnotation("ProductVersion", "8.0.7")
                 .HasAnnotation("Relational:MaxIdentifierLength", 63);
 
             NpgsqlModelBuilderExtensions.UseIdentityByDefaultColumns(modelBuilder);
@@ -36,8 +36,14 @@ namespace AzzyBot.Migrations
                     b.Property<int>("AzuraCastId")
                         .HasColumnType("integer");
 
+                    b.Property<DateTime>("LastUpdateCheck")
+                        .HasColumnType("timestamp with time zone");
+
                     b.Property<bool>("ServerStatus")
                         .HasColumnType("boolean");
+
+                    b.Property<int>("UpdateNotificationCounter")
+                        .HasColumnType("integer");
 
                     b.Property<bool>("Updates")
                         .HasColumnType("boolean");
@@ -69,7 +75,7 @@ namespace AzzyBot.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<int?>("GuildId")
+                    b.Property<int>("GuildId")
                         .HasColumnType("integer");
 
                     b.Property<decimal>("InstanceAdminRoleId")
@@ -233,7 +239,9 @@ namespace AzzyBot.Migrations
                 {
                     b.HasOne("AzzyBot.Database.Entities.GuildsEntity", "Guild")
                         .WithOne("AzuraCast")
-                        .HasForeignKey("AzzyBot.Database.Entities.AzuraCastEntity", "GuildId");
+                        .HasForeignKey("AzzyBot.Database.Entities.AzuraCastEntity", "GuildId")
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
 
                     b.Navigation("Guild");
                 });