+++ src/AzzyBot-Next/Utilities/Encryption/AesGcmCipher.cs
@@ -15,6 +15,7 @@ public sealed class AesGcmCipher(byte[] nonce, byte[] tag, byte[] cipher)
     public static AesGcmCipher FromBase64String(string data)
     {
         byte[] dataBytes = Convert.FromBase64String(data);
+
         return new(
             dataBytes.Take(AesGcm.NonceByteSizes.MaxSize).ToArray(),
             dataBytes[^AesGcm.TagByteSizes.MaxSize..],